pyproject.toml
[tool.poetry]
name = "grpc"
version = "0.1.0"
description = ""
authors = ["Ihor <cguilliman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
Django = "^4.2.5"
grpcio = "^1.58.0"
grpcio-tools = "^1.58.0"
loguru = "^0.7.2"
pydantic = "^2.4.0"
pymongo = "^4.5.0"
black = "^23.9.1"
isort = "^5.12.0"
icecream = "^2.1.3"
elasticsearch = "^8.10.0"
elasticsearch-dsl = "^8.9.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
